refactor: 1st iteration
