+++ integration/api/models/login.py
@@ -1,11 +1,6 @@
-from ...core.base_model import OrjsonBaseModel
+from pydantic import BaseModel
 
 
-class LoginRequest(OrjsonBaseModel):
+class LoginRequest(BaseModel):
     email: str
     password: str
-
-
-class Token(OrjsonBaseModel):
-    access_token: str
-    token_type: str
